backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.12
Pillow==12.0.0
pillow-heif==1.3.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
requests==2.32.3
boto3==1.35.92
